Move expandLinks into general-section
